requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
WTForms==3.1.0
Werkzeug==2.3.7
Pillow==10.0.1
qrcode==7.4.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
bcrypt==4.0.1
redis==5.0.1